review: address PR comments — cleaner ulong typing, drop dead code/comments
- SlotStore: _fileEra long+sentinel -> ulong? (era is ulong).
- StatelessBlockTree.Prefetch: length/i to ulong, drop (int)/(ulong)
  casts around BlockhashCache.MaxDepth.
- BlockValidator/BlockErrorMessages: itemCount ulong throughout, drop
  the asymmetric cast.
- ProcessingStats/BlockStatistics: BlockTo long -> ulong (matches
  BlockFrom), drop the (long)block.Number cast at construction.
- TxPoolTxSource: drop noisy "we have leftover candidates / no optimal
  picking needed" narration; behavior is in the code.
- UnclesValidator.IsKin: name the bounded int once, only widen to
  ulong where it interacts with header.Number.
- HeaderValidator: restore master's check order (parent → totalDiff →
  seal → gasUsed → gasLimit → timestamp → blockNumber) to avoid
  silently swapping which error a failing test sees first; drop the
  ValidateFieldLimit virtual stub — it has no overrides and just
  returns true.
- ProgressLogger: strip the explanatory ulong-vs-long comments; the
  -1 queue sentinel keeps a one-line note.

+++ src/Nethermind/Nethermind.Consensus/Processing/ProcessingStats.cs
@@ -25,7 +25,7 @@ namespace Nethermind.Consensus.Processing
     {
         public long BlockCount { get; set; }
         public ulong BlockFrom { get; set; }
-        public long BlockTo { get; set; }
+        public ulong BlockTo { get; set; }
         public double ProcessingMs { get; set; }
         public double SlotMs { get; set; }
         [JsonPropertyName("mgasPerSecond")]
@@ -486,7 +486,7 @@ namespace Nethermind.Consensus.Processing
             {
                 BlockCount = chunkBlocks,
                 BlockFrom = chunkFirstBlockNumber,
-                BlockTo = (long)block.Number,
+                BlockTo = block.Number,
 
                 ProcessingMs = chunkMs,
                 SlotMs = runMs,
